\name{lle_scurve_data}

\alias{lle_scurve_data}

\docType{data}

\title{ S-curve data  }

\description{ Syntheticly generated empiric data of a three-dimensional S-curve. }

\usage{ data(lle_scurve_data) }

\format{ matrix with dimensions \eqn{n=1000, m=3} }

\seealso{
  \code{\link{lle_scurve}}
}