\name{lle_spiral}

\alias{lle_spiral}

\title{ Example application on a n-dimensional spiral }

\description{ In this example nine three-dimensional spirals are generated. Every spiral has a higher number of windings than the previous. The first spiral's intrinsic dimension is determined as \eqn{m=1}. It can be observed that a higher number of windings with constant sampling rate leads to a higher calculated intrinsic dimension. Due to this fact, the last spiral's intrinsic dimension is computed as \eqn{m=2}.}

\usage{ lle_spiral() }