\name{lle_wave}

\alias{lle_wave}

\docType{data}

\title{ wavefile }

\description{ syntheticly generated wavefile with 10 modes. }

\usage{ data(lle_wave) }

\format{ numeric vector of length 17640 (0.4 seconds, 44 kHz) }

\details{ To analyse an acustic signal a wavefile with 10 modes was generated and transformed into text. }